backend/requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==43.0.3
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.4
typing_extensions==4.15.0
